sincronización master 2023-04-18

+++ src/main/java/epntech/cbdmq/pe/dominio/admin/MenuRol.java
@@ -18,8 +18,8 @@ public class MenuRol {
 
 	}
 
-	public MenuRol(Integer codMenu, Integer codRol, String permisos) {
-		this.menuRolId = new MenuRolId(codMenu, codRol);		
+	public MenuRol(Long codMenu, Long codRol, String permisos) {
+		this.menuRolId = new MenuRolId(codMenu, codRol);
 		this.permisos = permisos;
 	}
